requirements.txt
flask
authlib
python-dotenv
requests
openfga-sdk